requirements.txt
gpiod==2.3.0
gpiodevice==0.0.5
paho-mqtt==2.1.0
pms5003==1.0.1
pyserial==3.5
RPi.GPIO==0.7.1
